requirements.txt
Pillow ~= 7.1.2
numpy~=1.18.4
tensorflow-gpu~=2.2.0
scipy~=1.4.1
matplotlib~=3.2.1